feat: support SearchFilter
