+++ packages/form/demos/search-filter.tsx
@@ -3,16 +3,14 @@ import ProForm, { SearchFilter, ProFormText } from '../src';
 
 export default () => {
   return (
-    <div>
-      <SearchFilter onFinish={values => console.log(values)}>
-        <ProForm.Group title="分组一">
-          <ProFormText name="name" />
-        </ProForm.Group>
-        <ProForm.Group title="分组二">
-          <ProFormText name="name2" label="开始日期" />
-          <ProFormText name="name3" label="名称" />
-        </ProForm.Group>
-      </SearchFilter>
-    </div>
+    <SearchFilter onFinish={values => console.log(values)}>
+      <ProForm.Group title="分组一">
+        <ProFormText name="name" />
+      </ProForm.Group>
+      <ProForm.Group title="分组二">
+        <ProFormText name="name2" label="开始日期" />
+        <ProFormText name="name3" label="名称" />
+      </ProForm.Group>
+    </SearchFilter>
   );
 };
