+++ packages/form/src/BaseForm/index.tsx
@@ -4,6 +4,7 @@ import { FormProps } from 'antd/lib/form/Form';
 import { FormItemProps } from 'antd/lib/form';
 import FieldContext from '../FieldContext';
 import Submiter, { SubmiterProps } from '../components/Submiter';
+import { GroupProps } from '../interface';
 
 export interface CommonFormProps {
   submiterProps?: Omit<SubmiterProps, 'form'>;
@@ -15,7 +16,19 @@ export interface BaseFormProps extends FormProps, CommonFormProps {
     submiter: React.ReactNode,
   ) => React.ReactNode;
   fieldStyle?: React.CSSProperties;
-  formItemProps?: FormItemProps; // TODO
+  formItemProps?: FormItemProps;
+  groupProps?: GroupProps;
+}
+
+// TODO fix typescript
+export function createField(Field: any): typeof Field {
+  const FieldWithContext: React.FC<any> = (props: any) => {
+    const { fieldStyle, formItemProps } = React.useContext(FieldContext);
+    return (
+      <Field style={fieldStyle} formItemProps={formItemProps} {...props} />
+    );
+  };
+  return FieldWithContext;
 }
 
 const BaseForm: React.FC<BaseFormProps> = props => {
@@ -25,6 +38,7 @@ const BaseForm: React.FC<BaseFormProps> = props => {
     submiterProps,
     fieldStyle,
     formItemProps,
+    groupProps,
     ...rest
   } = props;
   const [form] = Form.useForm();
@@ -36,6 +50,7 @@ const BaseForm: React.FC<BaseFormProps> = props => {
       value={{
         fieldStyle,
         formItemProps,
+        groupProps,
       }}
     >
       <Form form={form} {...rest}>
