+++ packages/form/src/components/Checkbox/index.tsx
@@ -3,7 +3,7 @@ import { Form, Checkbox } from 'antd';
 import { FormItemProps } from 'antd/lib/form';
 import { CheckboxGroupProps, CheckboxProps } from 'antd/lib/checkbox';
 import CheckboxGroup from 'antd/lib/checkbox/Group';
-import { createField } from '../../FieldContext';
+import { createField } from '../../BaseForm';
 
 const Group: React.FC<FormItemProps &
   CheckboxGroupProps & {
