+++ packages/form/src/components/Group/index.tsx
@@ -1,16 +1,16 @@
 import React from 'react';
 import { Space } from 'antd';
 import { ConfigConsumer, ConfigConsumerProps } from 'antd/lib/config-provider';
+import FieldContext from '../../FieldContext';
+import { GroupProps } from '../../interface';
 import './index.less';
 
-export interface GroupProps {
-  title?: React.ReactNode;
-  style?: React.CSSProperties;
-  titleStyle?: React.CSSProperties;
-}
-
 const Group: React.FC<GroupProps> = props => {
-  const { children, style, title, titleStyle } = props;
+  const { groupProps } = React.useContext(FieldContext);
+  const { children, style, title, titleStyle, titleRender } = {
+    ...groupProps,
+    ...props,
+  };
   return (
     <ConfigConsumer>
       {({ getPrefixCls }: ConfigConsumerProps) => {
@@ -18,7 +18,7 @@ const Group: React.FC<GroupProps> = props => {
         return (
           <div style={style}>
             <div className={`${className}-title`} style={titleStyle}>
-              {title}
+              {titleRender ? titleRender(title) : title}
             </div>
             <Space size={32}>{children}</Space>
           </div>
