+++ packages/form/src/components/TimePicker/index.tsx
@@ -4,7 +4,7 @@ import ProField from '@ant-design/pro-field';
 import { Form } from 'antd';
 import { FormItemProps } from 'antd/lib/form';
 import { DatePickerProps } from 'antd/lib/date-picker';
-import { createField } from '../../FieldContext';
+import { createField } from '../../BaseForm';
 
 /**
  * 时间选择组件
