+++ packages/form/src/interface.d.ts
@@ -1,0 +1,6 @@
+export interface GroupProps {
+  title?: React.ReactNode;
+  style?: React.CSSProperties;
+  titleStyle?: React.CSSProperties;
+  titleRender?: (title: React.ReactNode) => React.ReactNode;
+}
